refactor: apply Feature-Sliced Design (FSD)
