+++ src/__tests__/localStorage.test.tsx
@@ -2,7 +2,7 @@ import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { beforeEach, expect, it } from 'vitest';
 
-import { SearchBar } from '../features/SearchBar.tsx';
+import { SearchBar } from '../features/search-character/ui/SearchBar.tsx';
 
 beforeEach(() => {
   localStorage.clear();
