+++ src/features/character-details/ui/DetailsView.tsx
@@ -2,10 +2,10 @@ import { type FC } from 'react';
 import { useSearchParams } from 'react-router-dom';
 
 import { CharacterDetails } from '@/entities/character/ui/CharacterDetails.tsx';
-import { CloseButton } from '@/shared/ui/CloseButton.tsx';
-import { ErrorMessage } from '@/shared/ui/ErrorMessage.tsx';
-import { OverlayWrapper } from '@/shared/ui/OverlayWrapper.tsx';
-import { Spinner } from '@/shared/ui/Spinner.tsx';
+import { CloseButton } from '@/shared/close-button/ui/CloseButton.tsx';
+import { ErrorMessage } from '@/shared/error-message/ui/ErrorMessage.tsx';
+import { OverlayWrapper } from '@/shared/overlay-wrapper/ui/OverlayWrapper.tsx';
+import { Spinner } from '@/shared/spinner/ui/Spinner.tsx';
 import { useGetCharacterByIdQuery } from '@/utils/api/rickAndMorty.ts';
 import { useLockBodyScrollOnMobile } from '@/utils/hooks/useLockBodyScrollOnMobile.ts';
