+++ src/features/search-character/ui/SearchBar.tsx
