+++ src/shared/close-button/ui/CloseButton.tsx
